pkgspec: allow empty package definitions
When a package is defined in yaml without attributes, we get a nasty
crash as we're getting None instead of a dict back from the yaml parser.
Therefore initializing with empty dict in that case.

This alone isn't veryusefuly yet, except for better diagnosis
(eg. will complain about missing autobuild-branch setting when
trying to do checkout), or when doing the checkout manually.

But later, we're going to support automatic imports (eg. from pypi)
and config templates, so we can reduce the config boilerplate, and
the machinery doesn't need explicit per-package settings for many
cases anymore.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>