deb_autopkg: tasks: pkg_build_target: fix regression in need_run checking
When checking whether a package needs to be (re)build, we first need the git
repo to be cloned. Doing so in PkgBuildTarget's need_run() method tries to
do so before the GitCloneTask subtask is run. It's an regression introduced
when we added multiple packagers (zypper vs apt) as dedicated subtasks and
ref head checking.

Fixing this by moving the statfile checking into the leaf nodes. This way,
all other subtasks (eg. git cloning) always get their chance to run.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>