deb_autopkg: add forced branch reset support
Add support for forced branch reset on each build.

When activated, the working copy is always forcefully reset to
the autobuild-ref before the actual build starts. Note that all
local changes to the current branch will be lost !

The feature is controlled by 'autobuild-force' package attribute
and disabled by default.

Signed-off-by: Enrico Weigelt, metux IT consult <info@metux.net>